requirements.txt
fastmcp
markdown
weasyprint
python-docx
